+++ src/button.h
@@ -3,19 +3,19 @@
 #include <cstdint>
 #include "config.h"
 
-uint8_t *keyScan();
+uint16_t *keyScan();
 /**
  * 清空按键标志位
  */
-void keyClear(uint8_t mask);
+void keyClear(uint16_t mask);
 /**
  * 检查并清空按键标志位
  */
-bool keyCheck(uint8_t mask);
+bool keyCheck(uint16_t mask);
 /**
  * 检查按键标志位
  */
-bool keyPeek(uint8_t mask);
+bool keyPeek(uint16_t mask);
 /**
  * 等待任意键
  */
